backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pyyaml==6.0.1
requests==2.31.0
huggingface_hub==0.16.4
diffusers==0.19.3
transformers==4.32.1
accelerate==0.21.0
# torch==2.0.1 - already installed, don't override
Pillow==10.1.0
python-multipart==0.0.6
safetensors==0.3.2
ftfy==6.1.3